fix(android): amend Ti.UI.Shortcut implementation
