+++ android/modules/ui/src/java/ti/modules/titanium/ui/UIModule.java
@@ -554,14 +554,4 @@ public class UIModule extends KrollModule
 			this.module.fireEvent(TiC.EVENT_USER_INTERFACE_STYLE, event);
 		}
 	}
-
-	/**
-	 * Manually define `createShortcut` to allow `ShortcutModule` to exist for 'click' events.
-	 * @return Shortcut proxy instance.
-	 */
-	@Kroll.method
-	public KrollProxy createShortcut()
-	{
-		return new ShortcutProxy();
-	}
 }
